Module aws_sdk_support::error

Modules

attachment_id_not_found

See AttachmentIdNotFound

attachment_limit_exceeded

See AttachmentLimitExceeded

attachment_set_expired

See AttachmentSetExpired

attachment_set_id_not_found

See AttachmentSetIdNotFound

attachment_set_size_limit_exceeded

See AttachmentSetSizeLimitExceeded

case_creation_limit_exceeded

See CaseCreationLimitExceeded

case_id_not_found

See CaseIdNotFound

describe_attachment_limit_exceeded

See DescribeAttachmentLimitExceeded

internal_server_error

See InternalServerError

Structs

AddAttachmentsToSetError

AddCommunicationToCaseError

AttachmentIdNotFound

An attachment with the specified ID could not be found.

AttachmentLimitExceeded

The limit for the number of attachment sets created in a short period of time has been

AttachmentSetExpired

The expiration time of the attachment set has passed. The set expires 1 hour after it

AttachmentSetIdNotFound

An attachment set with the specified ID could not be found.

AttachmentSetSizeLimitExceeded

A limit for the size of an attachment set has been exceeded. The limits are three

CaseCreationLimitExceeded

The case creation limit for the account has been exceeded.

CaseIdNotFound

The requested caseId couldn't be located.

CreateCaseError

DescribeAttachmentError

DescribeAttachmentLimitExceeded

The limit for the number of DescribeAttachment requests in a short

DescribeCasesError

DescribeCommunicationsError

DescribeServicesError

DescribeSeverityLevelsError

DescribeTrustedAdvisorCheckRefreshStatusesError

DescribeTrustedAdvisorCheckResultError

DescribeTrustedAdvisorCheckSummariesError

DescribeTrustedAdvisorChecksError

InternalServerError

An internal server error occurred.

RefreshTrustedAdvisorCheckError

ResolveCaseError

Enums

AddAttachmentsToSetErrorKind

AddCommunicationToCaseErrorKind

CreateCaseErrorKind

DescribeAttachmentErrorKind

DescribeCasesErrorKind

DescribeCommunicationsErrorKind

DescribeServicesErrorKind

DescribeSeverityLevelsErrorKind

DescribeTrustedAdvisorCheckRefreshStatusesErrorKind

DescribeTrustedAdvisorCheckResultErrorKind

DescribeTrustedAdvisorCheckSummariesErrorKind

DescribeTrustedAdvisorChecksErrorKind

RefreshTrustedAdvisorCheckErrorKind

ResolveCaseErrorKind