Module aws_sdk_support::model

Modules

attachment

See Attachment

attachment_details

See AttachmentDetails

case_details

See CaseDetails

category

See Category

communication

See Communication

recent_case_communications

See RecentCaseCommunications

service

See Service

severity_level

See SeverityLevel

trusted_advisor_category_specific_summary

See TrustedAdvisorCategorySpecificSummary

trusted_advisor_check_description

See TrustedAdvisorCheckDescription

trusted_advisor_check_refresh_status

See TrustedAdvisorCheckRefreshStatus

trusted_advisor_check_result

See TrustedAdvisorCheckResult

trusted_advisor_check_summary

See TrustedAdvisorCheckSummary

trusted_advisor_cost_optimizing_summary

See TrustedAdvisorCostOptimizingSummary

trusted_advisor_resource_detail

See TrustedAdvisorResourceDetail

trusted_advisor_resources_summary

See TrustedAdvisorResourcesSummary

Structs

Attachment

An attachment to a case communication. The attachment consists of the file name and

AttachmentDetails

The file name and ID of an attachment to a case communication. You can use the ID to

CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained

Category

A JSON-formatted name/value pair that represents the category name and category code

Communication

A communication associated with a support case. The communication consists of the

RecentCaseCommunications

The five most recent communications associated with the case.

Service

Information about an AWS service returned by the DescribeServices

SeverityLevel

A code and name pair that represents the severity level of a support case. The

TrustedAdvisorCategorySpecificSummary

The container for summary information that relates to the category of the Trusted Advisor check.

TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

TrustedAdvisorCheckRefreshStatus

The refresh status of a Trusted Advisor check.

TrustedAdvisorCheckResult

The results of a Trusted Advisor check returned by DescribeTrustedAdvisorCheckResult.

TrustedAdvisorCheckSummary

A summary of a Trusted Advisor check result, including the alert status, last refresh,

TrustedAdvisorCostOptimizingSummary

The estimated cost savings that might be realized if the recommended operations are

TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.