Module aws_sdk_support::operation

Structs

AddAttachmentsToSet

Adds one or more attachments to an attachment set.

AddCommunicationToCase

Adds additional customer communication to an AWS Support case. Use the caseId

CreateCase

Creates a case in the AWS Support Center. This operation is similar to how you create a case

DescribeAttachment

Returns the attachment that has the specified ID. Attachments can include screenshots,

DescribeCases

Returns a list of cases that you specify by passing one or more case IDs. You can use

DescribeCommunications

Returns communications and attachments for one or more support cases. Use the

DescribeServices

Returns the current list of AWS services and a list of service categories for each

DescribeSeverityLevels

Returns the list of severity levels that you can assign to a support case. The severity

DescribeTrustedAdvisorCheckRefreshStatuses

Returns the refresh status of the AWS Trusted Advisor checks that have the specified check

DescribeTrustedAdvisorCheckResult

Returns the results of the AWS Trusted Advisor check that has the specified check ID. You

DescribeTrustedAdvisorCheckSummaries

Returns the results for the AWS Trusted Advisor check summaries for the check IDs that you

DescribeTrustedAdvisorChecks

Returns information about all available AWS Trusted Advisor checks, including the name, ID,

RefreshTrustedAdvisorCheck

Refreshes the AWS Trusted Advisor check that you specify using the check ID. You can get the

ResolveCase

Resolves a support case. This operation takes a caseId and returns the