Module aws_sdk_support::output

Modules

add_attachments_to_set_output

See AddAttachmentsToSetOutput

add_communication_to_case_output

See AddCommunicationToCaseOutput

create_case_output

See CreateCaseOutput

describe_attachment_output

See DescribeAttachmentOutput

describe_cases_output

See DescribeCasesOutput

describe_communications_output

See DescribeCommunicationsOutput

describe_services_output

See DescribeServicesOutput

describe_severity_levels_output

See DescribeSeverityLevelsOutput

describe_trusted_advisor_check_refresh_statuses_output

See DescribeTrustedAdvisorCheckRefreshStatusesOutput

describe_trusted_advisor_check_result_output

See DescribeTrustedAdvisorCheckResultOutput

describe_trusted_advisor_check_summaries_output

See DescribeTrustedAdvisorCheckSummariesOutput

describe_trusted_advisor_checks_output

See DescribeTrustedAdvisorChecksOutput

refresh_trusted_advisor_check_output

See RefreshTrustedAdvisorCheckOutput

resolve_case_output

See ResolveCaseOutput

Structs

AddAttachmentsToSetOutput

The ID and expiry time of the attachment set returned by the AddAttachmentsToSet operation.

AddCommunicationToCaseOutput

The result of the AddCommunicationToCase operation.

CreateCaseOutput

The support case ID returned by a successful completion of the CreateCase operation.

DescribeAttachmentOutput

The content and file name of the attachment returned by the DescribeAttachment operation.

DescribeCasesOutput

Returns an array of CaseDetails

DescribeCommunicationsOutput

The communications returned by the DescribeCommunications

DescribeServicesOutput

The list of AWS services returned by the DescribeServices

DescribeSeverityLevelsOutput

The list of severity levels returned by the DescribeSeverityLevels

DescribeTrustedAdvisorCheckRefreshStatusesOutput

The statuses of the Trusted Advisor checks returned by the DescribeTrustedAdvisorCheckRefreshStatuses operation.

DescribeTrustedAdvisorCheckResultOutput

The result of the Trusted Advisor check returned by the DescribeTrustedAdvisorCheckResult operation.

DescribeTrustedAdvisorCheckSummariesOutput

The summaries of the Trusted Advisor checks returned by the DescribeTrustedAdvisorCheckSummaries operation.

DescribeTrustedAdvisorChecksOutput

Information about the Trusted Advisor checks returned by the DescribeTrustedAdvisorChecks operation.

RefreshTrustedAdvisorCheckOutput

The current refresh status of a Trusted Advisor check.

ResolveCaseOutput

The status of the case returned by the ResolveCase operation.